Recieving the error NameError at /assignment/1 undefined local variable or method `submission' for #<Student id: 2, name: "Heather Praisewater"> Did you mean? submissions submissions=
I searched and found that the grade.rb had :submission whereas all other programs used :submissions, plural. i changed them both to :submissions and ran it again and I get an identical error.

My question is: you said not to undo our changes, but if we don't get at least a different error message, can we assume the change we made was not what ruby wanted to work? in that case, would you suggest undoing that work, or going forward to see if you missed another one?